refactor runtime into its own crate
