+++ crates/par-builtin/src/builtin/debug.rs
@@ -1,9 +1,7 @@
 use std::sync::Arc;
 
-use par_core::{
-    frontend::{Definition, Module, Type, process},
-    runtime::Handle,
-};
+use par_core::frontend::{Definition, Module, Type, process};
+use par_runtime::readback::Handle;
 
 pub(super) fn external_module() -> Module<Arc<process::Expression<()>>> {
     Module {
