+++ crates/par-builtin/src/builtin/map.rs
@@ -3,10 +3,8 @@ use std::{collections::BTreeMap, future::Future};
 use crate::builtin::list::readback_list;
 use arcstr::literal;
 use num_bigint::BigInt;
-use par_core::{
-    frontend::{Definition, Module, Type, process},
-    runtime::Handle,
-};
+use par_core::frontend::{Definition, Module, Type, process};
+use par_runtime::readback::Handle;
 use std::sync::Arc;
 
 pub(super) fn external_module() -> Module<Arc<process::Expression<()>>> {
