+++ crates/par-builtin/src/builtin/nat.rs
@@ -2,10 +2,8 @@ use arcstr::literal;
 use num_bigint::BigInt;
 use std::{cmp::Ordering, sync::Arc};
 
-use par_core::{
-    frontend::{Definition, Module, ParString, Type, TypeDef, process},
-    runtime::Handle,
-};
+use par_core::frontend::{Definition, Module, ParString, Type, TypeDef, process};
+use par_runtime::readback::Handle;
 
 pub(super) fn external_module() -> Module<Arc<process::Expression<()>>> {
     Module {
