+++ crates/par-builtin/src/builtin/url.rs
@@ -4,10 +4,8 @@ use arcstr::literal;
 use percent_encoding::percent_decode_str;
 use url::Url as ParsedUrl;
 
-use par_core::{
-    frontend::{Definition, Module, ParString, Type, process},
-    runtime::Handle,
-};
+use par_core::frontend::{Definition, Module, ParString, Type, process};
+use par_runtime::readback::Handle;
 
 pub(super) fn external_module() -> Module<Arc<process::Expression<()>>> {
     Module {
