+++ crates/par-core/Cargo.toml
@@ -9,10 +9,9 @@ edition = "2024"
 playground = []
 
 [dependencies]
+par-runtime = { path = "../par-runtime" }
 futures = { version = "0.3.31", features = ["executor", "thread-pool"] }
 indexmap = "2.7.0"
-tokio = { version = "1.43.0", features = ["rt-multi-thread", "macros", "fs", "io-util", "io-std", "signal", "sync"] }
-stacker = "0.1.19"
 winnow = { version = "0.7.4", features = [] }
 miette = { version = "7.6.0", features = ["fancy"] }
 arcstr = "1.2.0"
