+++ crates/par-core/src/frontend_impl.rs
@@ -2,7 +2,6 @@ pub(crate) mod captures;
 pub(crate) mod language;
 pub(crate) mod lexer;
 pub(crate) mod parse;
-pub(crate) mod primitive;
 pub(crate) mod process;
 pub(crate) mod program;
 pub(crate) mod types;
