+++ crates/par-core/src/lib.rs
@@ -7,7 +7,7 @@ pub(crate) mod frontend_impl;
 pub(crate) mod location;
 #[path = "runtime/mod.rs"]
 pub(crate) mod runtime_impl;
-pub(crate) mod spawn;
 pub(crate) mod test_assertion;
+mod typed_readback;
 
 pub use facade::{execution, frontend, runtime, source, testing};
