+++ crates/par-runtime/src/lib.rs
@@ -1,0 +1,8 @@
+mod executor;
+pub mod fan_behavior;
+pub mod flat;
+pub mod primitive;
+pub mod readback;
+pub mod spawn;
+
+pub use executor::start_and_instantiate;
